Pugmas work

+++ src/main/java/me/pugabyte/nexus/models/MongoService.java
@@ -6,14 +6,21 @@ import me.pugabyte.nexus.Nexus;
 import me.pugabyte.nexus.framework.exceptions.BNException;
 import me.pugabyte.nexus.framework.persistence.MongoDBDatabase;
 import me.pugabyte.nexus.framework.persistence.MongoDBPersistence;
+import me.pugabyte.nexus.models.pugmas20.Pugmas20Service;
+import me.pugabyte.nexus.utils.PlayerUtils;
 import org.apache.commons.lang.Validate;
+import org.bukkit.Bukkit;
 import org.jetbrains.annotations.NotNull;
 
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+
+import static me.pugabyte.nexus.utils.StringUtils.isV4Uuid;
 
 public abstract class MongoService extends DatabaseService {
 	protected static Datastore database;
@@ -38,6 +45,8 @@ public abstract class MongoService extends DatabaseService {
 	@Override
 	@NotNull
 	public <T> T get(UUID uuid) {
+		if (this.getClass() == Pugmas20Service.class)
+			log("Get " + PlayerUtils.getPlayer(uuid).getName() + " [" + uuid.toString() + "]");
 //		if (isEnableCache())
 			return (T) getCache(uuid);
 //		else
@@ -55,8 +64,11 @@ public abstract class MongoService extends DatabaseService {
 
 	protected <T extends PlayerOwnedObject> T getNoCache(UUID uuid) {
 		Object object = database.createQuery(getPlayerClass()).field(_id).equal(uuid).first();
-		if (object == null)
+		if (object == null) {
+			if (this.getClass() == Pugmas20Service.class)
+				log("Creating new object " + PlayerUtils.getPlayer(uuid).getName() + " [" + uuid.toString() + "]");
 			object = createPlayerObject(uuid);
+		}
 		if (object == null)
 			Nexus.log("New instance of " + getPlayerClass().getSimpleName() + " is null");
 		return (T) object;
@@ -79,6 +91,10 @@ public abstract class MongoService extends DatabaseService {
 
 	@Override
 	public <T> void saveSync(T object) {
+		PlayerOwnedObject playerOwnedObject = (PlayerOwnedObject) object;
+		if (!isV4Uuid(playerOwnedObject.getUuid()))
+			return;
+
 		try {
 			database.merge(object);
 		} catch (UpdateException doesntExistYet) {
@@ -96,13 +112,33 @@ public abstract class MongoService extends DatabaseService {
 
 	@Override
 	public <T> void deleteSync(T object) {
+		PlayerOwnedObject playerOwnedObject = (PlayerOwnedObject) object;
+
+		if (!isV4Uuid(playerOwnedObject.getUuid()))
+			return;
+
 		database.delete(object);
-		getCache().remove(((PlayerOwnedObject) object).getUuid());
+		getCache().remove(playerOwnedObject.getUuid());
 	}
 
 	@Override
 	public void deleteAllSync() {
 		database.getCollection(getPlayerClass()).drop();
 		clearCache();
+	}
+
+	public void log(String name) {
+		try {
+			try {
+				throw new BNException("Stacktrace");
+			} catch (BNException ex) {
+				StringWriter sw = new StringWriter();
+				ex.printStackTrace(new PrintWriter(sw));
+				Bukkit.isPrimaryThread();
+				Nexus.fileLogSync("pugmas-db-debug", "[Primary thread: " + Bukkit.isPrimaryThread() + "] MongoDB Pugmas20 " + name + "\n" + sw.toString() + "\n");
+			}
+		} catch (Exception ex) {
+			ex.printStackTrace();
+		}
 	}
 }
